examples/cxl-mctp: fix fmapi set/add/release block size
Fixes GitHub Issue #19: cxl-mctp invalid extent size of init-add
command

The original code was not tested with DCD support in the kernel because
the necessary MCTP support patches were not rebased on it yet.
Without DCD kernel support, QEMU was hacked to directly add/release the
extents instead of using the pending list for this test code.

Now that we have kernel DCD support, the block size in should be changed from
an invalid size (128B) to something more sensible (128 MiB).

Polls device for total num extents after sending add/release commands to ensure
kernel and device finished processing. Without this check, a release request
can be sent before the extent it specifies is added, which causes it to fail
and vice versa for add requests.

The last test scenario in test_initiate_dc_release() is also updated.
The scenario tests adding 1 extent that is 3 blocks long and removing
the middle block. With the original QEMU hack that directly
added/released extents, the middle block would be released and result in
2 extents in the extent list, corresponding to the first and third
block.

However, with kernel DCD support, the kernel will release the entire
extent, resulting in an empty extent list. The test behavior doesn't
change, but this is reflected in the comment describing the expected
behavior.

Signed-off-by: Anisa Su <anisa.su@samsung.com>