lib/memdev: support security passthrough commands
Add support for Security Send (4600h) and Security Receive (4601h).

These are beyond the scope of the CXL specification and refer to
the Security Features for SCSI Commands (SFSC). For now only the
protocol 0x0 is supported which is the discovery[0] one, which a
max return of 256 protocols.

Unlike CXL, the NVME equivalent relies on explicit transport data
lengths, which will therefore always be more explicit to the user
than this implementation, but arbitrarily allowing 256 should be
enough.

[0]: https://www.seagate.com/content/dam/seagate/migrated-assets/staticfiles/support/docs/manual/Interface%20manuals/100293068m.pdf

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>