lib: Cleanup remaining ep upon cxlmi_free_ctx()
It is currently assumed that the user will manually close all
endpoints before freeing the context, and otherwise the leakage
will be the users fault. While this is valid, it does present
some asynchony respect to the scanning functionality, which
populates the endpoints on behalf of the user.

Change this to be a bit more defensive and free any loose
endpoints before freeing the context.

Signed-off-by: Davidlohr Bueso <dave@stgolabs.net>